+++ src/server.ts
@@ -66,6 +66,7 @@ connection.onInitialize((params): InitializeResult => {
 			},
 			executeCommandProvider: {
 				commands: [
+					CommandIds.LOWERCASE,
 					CommandIds.UPPERCASE,
 					CommandIds.EXTRA_ARGUMENT,
 					CommandIds.DIRECTIVE_TO_BACKSLASH,